add run-benchmarks.js for fast testing on windows
